Box entire value in a microstate
There was a lot of conflict surrounding whether an object was a
microstate, or whether an object was a microstate value. Sometimes a
microstate acted like the properties it wrapped, and sometimes it acted
as the state machine box that it is.

This enforces a clean separation between the microstate and its
value. All values in a microstate are now wrapped in contextualized
microstates, and the microstate forms an impenetrable box around its
properties. The only way to "unbox" the value of a microstate is to call
its `valueOf` method. E.g.

```js
let switchboard = new State({
  left: true,
  right: true
});

let leftOff = switchboard.left.set(false);
let rightOff = switchboard.right.set(false);

rightOff.valueOf() //=> { left: true, right: false }
rightOff.right.valueOf() //=> false.
```

In the cases where microstates are comprised of other microstates, it
will ensure boxing in the microstate constructor currently occupying a
particular slot.

```js
const Switch = new State({
  transitions: {
    toggle(current) {
      return !!current;
    }
  }
});

let board = new State({left: new Switch(false), right: new Switch(false)});
let leftOn = board.put({left: true});
let allOff = board.left.set(false);
```

It might be worth noting that to support this, the value exiting the box
vs the value entering the box are not the same unless it is a primitive
value. This is because the valueOf computation is lazy and will only
materialize the subtrees as they are requested.

```js
let object = {one: 'two'};
new State(object).valueOf() === object //=> false
```

We can return to this behavior, i.e. preserve the value if nothing has
changed, but it would mean a depth first traversal of the entire value
tree in order to search for changes, so I decided to err on the side of
computational laziness.