Isolate local tree passed into a transition.
The transition handler receives the structure tree at the point where
the transition is taking place. However, the path attribute of this
local tree still contains its context from the containing tree that it
was extracted out of. This "decontextualizes" that tree so that when
the transition handler sees it, it can be used just as though the tree
had been freshly analyzed, starting at that point.