Perform type initialization inside the tree analysis.
The fact that type shifting could occur after tree analysis was
complete threw a monkey wrench into parametric types.

This makes tree analysis happen all in once fell swoop, since it turns
out that we don't need to actually query the state at the node. We
only need to invoke the constructor to see if there is any type
initalization.

Since this doesn't need to re-construct types from values, the
`getType` method was removed in lieu of the `toType` function which
only considers Constructors.