Introducing type parameters for Array and Object. (#70)
Part of what makes microstates so powerful is their composability. We
can take the data and operations of one class and "place" them into
another class. Sometimes however, the nature of a type is to manage or
"contain" another type. In other words, it doesn't have much meaning
in relation to the type. Arrays and Objects are examples of this. They
should be able to work with _any_ type, and still be able to leverage
the transitions of that type. That's not currently the case.

As a concrete example, before in order to make a todo list, all
of the transitions had to live on the "TodoList" class because the
array only holds "data".

```js
class TodoList {
  items = Array

  setTodoDescription(todo, description) {}

  markDone(todo) {}
}
```

But the transitions associated with an individual item, are really
about that item. What we'd really like to be able to do is to have a
class:

```js
class TodoItem {
  description = String;
  isDone = Boolean;
}
```

and then be able to use the string and boolean transitions in there.

With parameterized types this becomes a possibility. Now, we can
"embed" the `TodoItem` inside the array by declaring our TodoList as:

```js
import { parameterized } from 'microstates';

class TodoList {
  items = parameterized(Array, TodoItem)
}
```