Introducing `static create` method for initialization (#73)
When modeling state machines, it's necessary to be able to initialize into a particular
state based on value. Previously, this was accomplished by returning a microstate from the constructor. This lead to exceeds maximum call stack when destination type extended from type shifting class. 

This PR introduces a new convention which allows the developer to describe how a state should be initialized for a specific type. `create` method must return a microstate that represents the state that you want the state to initialize into or nothing.  The create method takes one argument which is any `value`.

```js
import { create } from 'microstates';

class Session {
  content = null;

  static create(session) {
    if (session) {
      return create(AuthenticatedSession, session);
    } else {
      return create(AnonymousSession);
    }
  }
}

class AuthenticatedSession extends Session {
  isAuthenticated = true;
  content = Object;

  logout() {
    return create(AnonymousSession);
  }
}

class AnonymousSession extends Session {
  content = null;
  isAuthenticated = false;

  authenticate(user) {
    return create(AuthenticatedSession, { content: user });
  }
}
```

The create method follows the same semantics as the constructor. When you return a microstate, it'll use the type and value in the microstate for `this` node. If you return anything else, it'll ignore the result and create an empty instance of that type. 

One perk of this solution is that all types that extend from the root type will automatically get this create method.