Fix Array transitions
With the advent of parameterized types, the tree can extend _below_
the Array type, so it isn't enough just to work with the state or the
value. We need to keep the tree in sync.

This implements a `splice` transition on the `ArrayType` which
actually modifies the microstate tree to keep in sync with the
value, and then `push`, `pop`, and `filter`, are implemented in terms
of `splice`. Any nodes that can be shared are shared.

This is clearly not optimal, but we need to see what tree modification
patterns look like for Object and others before abstracting.