WIP - Associate a value and state with every node in the structure
There are two types of nodes in the tree, `PrimaryValue` nodes, and
`NestedValue` nodes. Primary Value nodes have their own value, and
they don't derive them from anywhere in the tree. NestedValue nodes
have a `root` value which the then derive their value by lensing them
out of the root.

Big changes are that now, when creating a transition context, we need
to update the root value of all of the nested nodes to be the local
value at the transition site, and then once the transition is
complete, update all the nested nodes to have their root be the new
global value.