Concentrate Lens logic in the lens module. (#100)
We want to get to a point where we merge nodes and trees so that we can calculate complete things like states from directly within the node without any further context.

As a first step towards making that work, this removes lensing logic where it isn't needed, and moves ones that are still needed into the lensing module. That way, when we do need to make changes to the way the tree is changed, it can happen all in a single place.