Introduce general Array transformation function.
It looks like the only things actually needed to transform a current
array is a list of the children and a type parameter. I think this
might be generalizable into the object type as well.

The array unshift functionality has been removed in order to maintain
a uniform type signature.