Fully generic tree mapping.
Before we were doing a separate tree mapping for updating the metadata
and for creating the identity. This unifies this under a single
typeclass `Tree`. This is important because ultimately queries will
not be microstates, but they _will_ be trees.

The basic idea is that certain objects are container objects, or
"trees". Anything that implements the Tree typeclass can have its
structure replicated via `Tree.map`, but the actual objects that that
sit at each location can be completely different.

So in the case of context mapping, we're mapping a microstate tree to
a microstate tree, but in the case of identity mapping, we're mapping
a microstate tree to an identity tree.

With the introduction of queries, we'll be mapping a mixed microstate /
query tree to a mixed identity / query tree.