Bring `from()` back from the dead.
The previous version of microstates had a `from` method wherein you
could bring to life an entire microstate tree just by passing in a
JavaScript object literal.

The type of the corresponding microstate tree is derived from the
JavaScript literal at each point.

This just uses the initialization transition to determine which type
the microstate should _actually_ be, and then creates it.