Make Tree.map() pass in the path of the node being mapped.
Previously when constructing an identity, we were using the path that
was stored on the microstate metadata. However, when mapping over
queries, the query object does not have meta, and so is not aware of
its own path. However, it is trivial to derive this during the tree
map.

In prepration for other objects than microstates being in tree (Query
objects), this uses the path from the tree map rather than the
microstate.