Create a hash based dictionary object to use as an identity map.
When mapping an identity tree, we want to use the same identity for
each microstate no matter _where_ it appears in the tree. To use the
TodoMVC example, if a `Todo` item appears in both the `todos` list and
also the `active` list, then it should use the exact same Id proxy to
represent the object.

In order to accomplish this, we have to be able to lookup objects
based on what makes them unique. This introduces a `Hash` typeclass
which lets you uniquely identify any object based on a set of unique
properties. It has a single method `digest` which takes an instance of
the Hash class, and returns an _Array_ of primitive objects that
uniquely identify it. In this way, we can take to microstates that
have the same Type and state, and map them back to the same identity
object. Or, take two Filters, that have the same microstates, and
predicate and map them back to the same identity.