Always use a context-free microstate when lens setting.
The `microstate.set()` method ultimately is responsible for producing
a new microstate and then embedding it into the same location using
the implicit lens. This microstate that gets passed to the lens must
be "context-free" or sitting at the head of its own tree.

There is an optimization however to cover the case where the value
being passed in is the same as the current microstate's state. In that
event, we were attempting to re-use the current microstate. However,
what we actually wanted was to return the _decontextualized_ version
of the current microstate.

This change makes sure to always use the decontextualized microstate
before setting it with the lens.

As a bonus, this means that you can now "pass" microstates from one
tree to another and it will work just fine.

```js
class A {
  b = class B {
    c = class C {
      value = Number;
    }
  }
}

let a1 = create(A, {b: {c: { value: 42 }}})
let a2 = create(A);

a2.b.c.set(a1.b.c); //no problem!

a2.b.c.value.state; //=> 42
```