Make DSL expansion 100% deterministic.
The "sugar" package allowed for any module to register its own
syntactic transformations. This prevented circular dependiencies and
added a theoretical advantage in that the microstates syntax could be
extended. While this _might_ be something we want in the future, it
also introduced some non-determinism since the order in which
syntactic transformations were registered resulted in different
outcomes.

This completely rewrites the syntactic transformation system to be
100% deterministic and located in a single place. (dsl.js)

It's very clear what the precedence of rules are, and what _all_ the
rules are just by looking in a single place.

In order to test that syntactic sugar is working, many of the
testcases have been converted to use it rather than instantiate the
types directly.