Allow consumers to use the Identity / Store API directly. (#194)
Previously, there was no way to use the idenity mechanism directly,
and the only way to make reactive systems with microstates (without
rolling your own) was to use `Observable` integartion.

This exports the `Identity` mechansim on which Observable is
based. It is exported as `Store`.

```
import { Store } from 'microstates';

let store = Store(create(Number, 42), microstate => {
    console.log(microstate)
});
```

Previously, the result of the callback was used as the return value of
the store function, but this didn't really make sense. Now, the
callback is ignored.