Make transitions functions stable per location.
Transition closures were bound to an instance. That way, they could be
separated from the instance and passed around as independent
actions. In terms of value stability, that means that
if an instance changed, then the transition functions attached to it
also needed to change and any of the actions passed around as props
would trigger re-renders.

In reality though, if I want to pass `bool.toggle` as an action, the
`toggle` function really doesn't need to change because it operates
on a *location* in the microstate tree. The actual microstate that
inhabits that location changes with each transition, but the functions
associated with that location do not.

This makes the transitions _super_ stable by storing the path
information in the Id class instead of the Id instance. Previously,
transition dispatch looked like:

1. get the path off of the instance
2. invoke the transition based off looking up the microstate at that
location.

This has been changed to:

1. get the path off a class parameter.
2. invoke the transition based off looking up the microstate at that
location.

With path now being implicity to the class, that means all instances
of that class share their methods on the class prototype, and once
again stability is introduced using the vanilla JavaScript way.