Add benchmarks to set bar for performance improvements.
We're seeing some pretty atrocious performance charateristics on
certain microstates, and so we need a way to frame the problem, point
at it, and also be able to measure what success would look like in the
event that we fix it.

This adds a benchmark suite to microstates that can be invoked with

```
yarn bench
```

It outputs its data in tabular form, but it would be awesome if we
could track the output over time and only accept pull requests that
don't downgrade performance by a significant margin.