Upgrade Babel to 7.0, chechkin yarn.lock
With the public release of babel, it broke our build.

This uprades our Babel and also checks in our yarn.lock so that future
releases don't break our build. Since we're not yet testing with a
bunch of different libraries, this ought to be ok. We don't have many
external dependencies anyway.

What was really broken by this was our build test and release process,
not the library itself, so that's what we're locking down with
yarn (our dev tools used to evolve microstates).