Add benchmarks for TodoMVC.
== Purpose

We've been working on a bunch of performance fixes by making the
properties of all microstates (normal, object, and array) lazy. We've
also been making the store lazy as a performance booster. We need a
way to measure the improvements against how they actually would look
in a microstates application though and not just create empty stores
or meaningless arrays. We need a benchmark that simulates something
approaching a real usage.

== Approach

The TodoMVC app, which is already used in our tests is very close to
what you would see in an application. It has an item model, and a list
model that tracks those items.

In order to simulate production usage, there are two cases that we
want to capture: initialization, and addition. This is because every
application has to start with an beginning state. It cannot be
prohibitively expensive to initialize a store with existing
application state. Also, it cannot be prohibitively expensive to
add items or take them away because that's the UX right there. If it's
slow, then it's a no-go.

These benchmarks highlight the current performance problems so that
any solution can be measured against them.

┌────────────┬─────────────────────────────────────────────────────────┬───────────┬──────────┬─────────────┐
│ Suite      │ Benchmark                                               │ ops/sec   │ σ        │ sample size │
├────────────┼─────────────────────────────────────────────────────────┼───────────┼──────────┼─────────────┤
│ TodoMVC    │ Store(create(TodoMVC, MUCH_TODO)); //100 todos          │ 5.38      │ ± 3.15 % │ 19          │
├────────────┼─────────────────────────────────────────────────────────┼───────────┼──────────┼─────────────┤
│ TodoMVC    │ store.todos.push(ONE_MORE_TODO); //add to a list of 100 │ 36.64     │ ± 2.32 % │ 49          │
└────────────┴─────────────────────────────────────────────────────────┴───────────┴──────────┴─────────────┘

== Implementation notes.

We picked 100 as the number todos that we tested against. We
actually tried more than that, but the benchmarks never could even
complete a single revolution.

Going lazy is not just for performance. It will allow us to have
recursive and circular microstates. E.g.

```js
class Tree {
  content = Object;
  children = [Tree];
}
```

== Further Exploration

The total suite of benchmarks that run here takes upwards of 3 minutes
to run. We could think about removing some of them as they seem
redundant.