Make all microstate and store properties lazy.
We were experiencing some performance bottlenecks with Arrays, which
made microstates impractical for certain use-cases. This was because
we were doing several things. The first was that state for a parent
was being eagerly constructed based on the child states.

So, for example, if I had a class

```js
class Person {
  firstName = String;
  lastName = String;
}
```

When we called `create(Person, undefined)`, it would loop over and do a `set()`
on each property, which would then integrate the default string state
into the `firstName` and `lastName` properties.

So as a result, `create(Person, undefined).state` would be
`{firstName: '', lastName: ''}`. This also, includes cases beyond just the
default state where there might be complex initialization logic.

The problem is that it means in order to determine the top level
state, you have to examine all of the constituent states, and for each
constituent state examine all of _its_ constituent states. So
basically, you have to examine the tree in its entirety when you make
a change.

The other problem is that it makes self-referential microstates
impossible:

```js
class Tree {
  children = [Tree];
}
```

Because in order to find out the state of the containing tree, you
have to know the state of the inner trees, each of which needs to know
the states of all _its_ inner trees.

This change allows for computing substates to be lazy by relaxing that
constraint. Now, a microstate closes over _only_ the state that you
initially give it upon `create`, and it can diverge from being purely
the sum of its substates. This lets us make the substates lazy,
computed properties, and it means that we don't have to reify the
entire tree on every `create()`. That means significant performance
gains.