Introducing Femtostates.
What began as performance work back in August to enable modeling
BigTest state with microstates ends here with a more complete
restructuring of the codebase to support performance.

The problem, as outlined in the summaries for
https://github.com/microstates/microstates.js/pull/209 and
https://github.com/microstates/microstates.js/pull/222 is that too
much eager computation was happening while setting up array and object
microstates. Also, too much computation was happening for normal
object construction since the properties were being computed eagerly
and it would have bitten us as trees got larger. Really, the only
reason Arrays blew up fast was because they represent big trees.

This change expands and encompasses all of the work done in those two
pull requests and supercedes them. Specifically, there are two major
changes in behavior, and a number of API changes needed to support them.

== Behavior Changes

1. All microstates are now completely lazy. This includes most
microstates, including parameterized Objects.

Every property in every microstate is lazily mounted, so if I say

```js
let person = create(class Person {
  car = create(class Auto {
    constructor() { console.log('constructing auto!!!')}
  })
})
// logs nothing

person.auto
// logs 'constructing auto!!!'

2. Array microstates can no longer be accessed by index.

We were having to eagerly define a property descriptor for each member
of an array. so we do a `defineProperty` for `0`, `defineProperty` for
`1`. As arrays grow to 1000 elements, this became untenable. Without
the ability to use `Proxy`, there's no way to make this lazy.

Instead, we make Array fully support the enumerable API so that it can
be enumerated, but not accessed directly. This prevents actually
instantiating any substates until the actual point at which the
iterator passes over it.

before:

```js
let array = create([], [1, 2, 3])
array[0].state //=> 1
array[2].state //=> 1
```

after:
```js
let array = create([], [1, 2, 3])
for (item of array) {
  console.log(item.state)
}
// logs 1 2 3
```

If you absolutely must have indexed access, you can "slurp" the array
microstate into a javascript array with destructuring assignement:

```js
let [...slurp] = create([], 1, 2, 3);
slurp[0].state //=> 1
slurp[1].state //=> 2
```

But of course that will eagerly compute the entire contents of the
array :)

3. As a result of all these changes, recursively defined microstates
now work!

```js
class Tree {
  name = String;
  children = [Tree];
}
```

is a perfectly valid definition of a microstate.

== API Changes

In order to support this extreme laziness, some parts of the API had
to change. These are the breaking changes:

1. there is no longer a `.state` property on _every_ microstate. The
requirement for the state to be _the_ canonical reference for the
value at that point, and also to be in sync with all the `.state`
properties down the tree was one of the main culprits of the
performance problems. In order to compute it we had to walk the entire
tree. It just had to go.

Instead, leaf microstates have the alias of `.state` to be their
value. The value can be accessed via the `valueOf` method. There's
nothing to stop you from implementing a `state` property as a computed
on your microstate Type, but what it will contain is up to you.

2. valueOf()

Now that `.state` is not available from most microstates, only those
that contain concrete values, you need a way to retrieve the simple,
POJO value associated with that point. This is where they `valueOf()`
method comes in. It will retreive the value of the total value at the
point of the microstate.

3. Severe restrictions on type-shifting.

As part of the the move to complete laziness, it means that the
microstate tree needs to be a pure function of the serialized (POJO)
state. This must be so because in lazy-land, the root microstate
conceptually "closes" over the root value from scratch with every
single transition, and in that way the tree is incrementally
constructed from the top down every time.

This means that there can't be anything stored in the microstate tree
that isn't also stored in the pojo tree. In practice I don't think
it'll be a problem, and it  doesn't mean that we can't have
polymorphism, it just has to be more structured. Basically, have a
single class "dispatches" to the actual runtime class based on
information contained in the state. This is in effect what the
`literal` class does. It looks at the native value and initializes to
the corresponding microstate value.

We could also introduce a specific helper to make this easier:

```js
const Maybe = Union(T => ({
  just: class Just {
    isJust = true;
    content = T;
  },
  nothing: class Nothing {
    isNothing = true;
  }
}))
```

Which would generate a `Maybe` class, that would swap out to the
actual value based on what it saw.

== Implementation Notes

1. In order to do the lazy treemapping and stable id generation, we
have a new typeclass `Profunctor` which is based on the Profunctors
I've seen in the world.
2. Queries now work on any javascript iterable, and are based on the
experimental generation.js https://github.com/cowboyd/generation.js We
have proofs of concept

== Open Questions

1. The iterators are very cool and enable the move to total laziness,
but it's very, very verbose.
2. We could probably move over to a standard lens library, since we
only use them again now for basic datatypes object and array.
3. Since Queries are just operating on enumerable, do we even ship
them with microstates now, and instead just use the separate query
library?