Rework opening line.
There's a lot of buzzword in the opening line of the readme that might
scare folks away.

Now that I've had a lot of conversations with audience members
at React ATX, Ember ATX, and now Framework summit. I believe the most
compelling arguments for Microstates are actually pretty simple.

1. Benefits of immutability, without the overhead (Think mutably, act immutably.)
2. performance via stable objects.
3. dead simple re-use via composition.

The opener talks a lot about the mechanics, but I worry that that can
obfuscate those key benefits.

This addresses those benefits head-on, letting the rest of the README
take care of the how.

Alternatives
===

- Microstates let you reap the benefits of strict immutability, yet
keep the simplicity of the standard JavaScript object model.
- Microstates: pure functions over immutable data doesn't have to be
hard.