Update documentation for missing state.
With the advent of the femtostates architecture, the `state` property
is no longer on composite microstates (those microstates which are
made out of combining other microstates).

Now, the only `.state` property is on primitive microstates (Number,
String, Any, Constant)

This updates the docs to use the `valueOf` which is the new preferred
method for extracting the aggregate value out of a microstate.

Also, Array microstates no longer allow direct, keyed access. Instead,
the docs now use iteration for the demos.