Bring back observable integration.
As part of the femtostates refactor, support for `Symbol.observable`
was removed. This brings it back.

There was talk about whether to put it directly on the `store` to
support cursors, but for the moment, I think it wise to just not
change to much at one time.

This maintains the original API 100%

When we have a nice think about how we want to maintain observable
pointers deep into the tree and how that will work, then we can talk
about breaking the API. Until then, let's keep it as a variable that
doesn't move.