Remove fully polymorphic lenses.
We were paying the cost for lens polymorphism (multiple kinds of
container objects) when in reality, the only ones we were using were
the native JavaScript Array and Object lenses.

What's cool about this is that when constructing a lens, you can also
pass the _type_ of container object the lens is for, that way, it will
preserve how to do the actual setting of a new value at the location
without actually having to remember the object. This let's you walk a
tree and record how it was constructed without actuall recording the
tree itself.

The idea comes from purescript lenses which have the concept of
polymorphic containers:
https://pursuit.purescript.org/packages/purescript-profunctor-lenses/3.8.0/docs/Data.Lens.At

In practical terms if I have an array lens, I can say:

```js
let arrayLens = At(0, []);
let objectLens = At(0, {});
```

And then the array-fullness is stored in the lens itself, without
actually storing the original array.

```js
set(arrayLens, 5, undefined) //=> [5]
set(objectLens, 5, undefined) //=> {'0': 5}
```

While useful, the only shapes we accounted for were simple Arrays and
Objects. Hardly worth the overhead both in terms of code-size and
computation.

This change removes that general polymorphism in favor of a simple,
inline polymorphism: we just check if we have an array or an object
when making decisions.

The upshot of this is that the interface of the `At` method remains
the same, so if we want to be more generally polymorphic in the
future, no code need change.