Expose raw test command.
Purpose
==

When we added coverage, it took over the test command so that coverage
was being run on every since test run. However, when developing it's
often very desirable to _not_ run coverage, only the test suite for
two reasons:

1. The test suite without coverage takes < 1s to run on a 2018 macbook
pro, whereas it takes > 3s to run with coverage.
2. It appends coverage data to the output no matter what, which can
make it difficult to interpret focused tests.

Approach
==

This splints the test command into two commands "test" and
"coverage". The "test" command is as it was before: just running the
test suite. The "coverage" command runs the test command but with
coverage, and then the "coverage" command becomes "coveralls" which
actually posts the results of the coverage report.

Finally, to simplify the build commands, we switch to using yarn on
travis since it requires less configuration out of the box.